requirements.txt
appium-python-client
selenium
pytest
webdriver-manager

